+++ bytebank/src/main/kotlin/modelo/Analista.kt
@@ -1,3 +1,5 @@
+package modelo
+
 class Analista(
     nome: String,
     cpf: String,
@@ -6,7 +8,7 @@ class Analista(
 
     override val bonificacao: Double
         get() {
-            println("Bonificação Analista")
+            println("Bonificação modelo.Analista")
             return salario * 0.1
         }
 }
