+++ bytebank/src/main/kotlin/modelo/CalculadoraBonificacao.kt
@@ -1,3 +1,5 @@
+package modelo
+
 class CalculadoraBonificacao {
 
     var total: Double = 0.0
@@ -8,11 +10,11 @@ class CalculadoraBonificacao {
         this.total += funcionario.bonificacao
     }
 
-//    fun registra(gerente: Gerente) {
+//    fun registra(gerente: modelo.Gerente) {
 //        this.total += gerente.bonificacao
 //    }
 //
-//    fun registra(diretor: Diretor) {
+//    fun registra(diretor: modelo.Diretor) {
 //        this.total += diretor.bonificacao
 //    }
 }
