+++ bytebank/src/main/kotlin/modelo/ContaCorrente.kt
@@ -1,0 +1,16 @@
+package modelo
+
+class ContaCorrente(
+    titular: String,
+    numero: Int
+) : Conta(
+    titular = titular,
+    numero = numero
+) {
+    override fun sacar(valor: Double) {
+        val valorComTaxa = valor + 0.1
+        if (this.saldo >= valorComTaxa) {
+            this.saldo -= valorComTaxa
+        }
+    }
+}
