+++ bytebank/src/main/kotlin/modelo/SistemaInterno.kt
@@ -1,3 +1,5 @@
+package modelo
+
 class SistemaInterno {
 
     fun entra(admin: Autenticavel, senha: Int) {
