+++ bytebank/src/main/kotlin/teste/testaCopiasEReferencias.kt
@@ -1,3 +1,6 @@
+import modelo.ContaCorrente
+import modelo.ContaPoupanca
+
 fun testaCopiasEReferencias() {
     val numeroX = 10
     var numeroY = numeroX
@@ -11,8 +14,8 @@ fun testaCopiasEReferencias() {
     val contaJoao = ContaCorrente("João", 1002)
     var contaMaria = ContaPoupanca("Maria", 1003)
 
-    println("Titular Conta João: ${contaJoao.titular}")
-    println("Titular Conta Maria: ${contaMaria.titular}")
+    println("Titular modelo.Conta João: ${contaJoao.titular}")
+    println("Titular modelo.Conta Maria: ${contaMaria.titular}")
 
     println(contaJoao)
     println(contaMaria)
