+++ bytebank/src/main/kotlin/teste/testaFuncionarios.kt
@@ -1,3 +1,8 @@
+import modelo.Analista
+import modelo.CalculadoraBonificacao
+import modelo.Diretor
+import modelo.Gerente
+
 fun testaFuncionarios() {
     val alex = Analista(nome = "Alex", cpf = "111.111.111-11", salario = 1000.0)
 
@@ -14,7 +19,7 @@ fun testaFuncionarios() {
     println("Bonificacao da ${fran.nome}: ${fran.bonificacao}")
 
     if (fran.autentica(senha = 1234)) {
-        println("Gerente autenticado com sucesso!")
+        println("modelo.Gerente autenticado com sucesso!")
     } else {
         println("Falha na autenticação!")
     }
@@ -28,7 +33,7 @@ fun testaFuncionarios() {
     println("PLR: ${gui.plr}")
 
     if (gui.autentica(senha = 4000)) {
-        println("Gerente autenticado com sucesso!")
+        println("modelo.Gerente autenticado com sucesso!")
     } else {
         println("Falha na autenticação!")
     }
